uop: move mop_cleanup to symbolic, breaking intra-uop cycle
